creating cakemail templte
This commit should create a folder named cakemail-template which will be compressed to a zip file in order for cakemailto accept it.

There should also appear two other html files, one with the style tags included in the header (to be inlined by mailchimps inlener tool) and one with all of the styles inline.